add permission limits and fix bug of creating/updaating user password

+++ user_profile/api/views.py
@@ -5,11 +5,13 @@ from user_profile.api.seriaiizers import ProfileSerializer, UserSerializer
 
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.parsers import JSONParser
 from rest_framework.authtoken.models import Token
 
 
+@permission_classes((IsAuthenticated,))
 @api_view(['GET'])
 def api_detail_user_view(request, id):
     try:
@@ -28,6 +30,7 @@ def api_detail_user_view(request, id):
         return Response(message)
 
 
+@permission_classes((IsAuthenticated,))
 @api_view(['PUT'])
 def api_update_user_view(request, id):
     try:
@@ -36,12 +39,19 @@ def api_update_user_view(request, id):
     except User.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+    authenticated_user = request.user
+    if user != authenticated_user:
+        return Response({'Response': "You don't have permission to edit it."})
+
     if request.method == 'PUT':
         data = JSONParser().parse(request)
         user_serializer = UserSerializer(user, data=data)
         profile_serializer = ProfileSerializer(profile, data=data)
         if user_serializer.is_valid():
-            user_serializer.save()
+            user = user_serializer.save()
+            if data['password']:
+                user.set_password(data.get('password'))
+                user.save()
         else:
             return Response(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -53,6 +63,7 @@ def api_update_user_view(request, id):
         return Response({"success": "update successful"}, status=status.HTTP_202_ACCEPTED)
 
 
+@permission_classes((IsAuthenticated,))
 @api_view(['DELETE'])
 def api_delete_user_view(request, id):
     try:
@@ -61,14 +72,19 @@ def api_delete_user_view(request, id):
     except User.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+    authenticated_user = request.user
+    if user != authenticated_user:
+        return Response({'Response': "You don't have permission to delete it."})
+
     if request.method == 'DELETE':
-        operation = user.delete()  # 會一併刪除profile
+        operation = user.delete()  # 刪除user會一併刪除profile
         if operation:
             return Response({"success": "delete successful"})
         else:
             return Response({"failure": "delete failed"})
 
 
+@permission_classes((AllowAny,))
 @api_view(['POST'])
 def api_create_user_view(request):
     if request.method == 'POST':
@@ -77,6 +93,8 @@ def api_create_user_view(request):
         user_serializer = UserSerializer(data=data)
         if user_serializer.is_valid():
             new_user = user_serializer.save()
+            new_user.set_password(data.get('password'))
+            new_user.save()
         else:
             return Response(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -97,14 +115,15 @@ def api_create_user_view(request):
         }
         return Response(message, status=status.HTTP_201_CREATED)
 
+
 @api_view(['POST'])
 def obtain_auth_token(request):
     if request.method == 'POST':
         data = JSONParser().parse(request)
         user = authenticate(
-            username=data.get('username', ''), 
+            username=data.get('username', ''),
             password=data.get('password', '')
-            )
+        )
         print(user)
         if user is not None:
             token = Token.objects.get(user=user).key
